domain-web/jbb-frontend-web/src/main/java/org/jbb/frontend/web/faq/controller/FaqController.java
/*
 * Copyright (C) 2017 the original author or authors.
 *
 * This file is part of jBB Application Project.
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  You may obtain a copy of the License at
 *        http://www.apache.org/licenses/LICENSE-2.0
 */

package org.jbb.frontend.web.faq.controller;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;

@Controller
public class FaqController {

    private static final String FAQ_VIEW_NAME = "faq";

    @RequestMapping("/faq")
    public String getFaq() {
        return FAQ_VIEW_NAME; //NOSONAR
    }


}
